requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
certifi==2022.6.15
chardet==4.0.0
charset-normalizer==2.0.12
discord.py==1.7.3
idna==3.3
multidict==6.0.2
pip==21.3.1
praw==7.6.0
prawcore==2.3.0
python-dotenv==0.20.0
requests==2.28.0
setuptools==60.2.0
typing-extensions==4.2.0
update-checker==0.18.0
urllib3==1.26.9
wcwidth==0.2.5
websocket-client==1.3.2
wheel==0.37.1
yarl==1.7.2
